Move fetching the announcements up to SideBar
